% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatio.R
\name{calcRatio}
\alias{calcRatio}
\title{Calculate read ratio and smoothed read ratio.
　
Calculate read ratio at each interval based on normalized read counts
obtained via \code{\link[=doSInG]{doSInG()}}. Read ratios calculated at each interval
usually contain data points with stochastic error which are
represented as unexpected changes in read ratio between adjacent
intervals. To remove those errorneous intervals, this function
also calculate average read ratio within the specified window.}
\usage{
calcRatio(object, window = 10000, min_marker = 0)
}
\arguments{
\item{object}{A \code{SInG} object after running \code{\link[=doSInG]{doSInG()}}.}

\item{window}{A integer value to specify the window size in base pairs.}

\item{min_marker}{A integer to indicate the minimum number of markers
with in each window. If the number of markers within a window with the
size \code{window} is less than \code{min_marker}, use \code{round(min_markers/2)}
markers at upstream and downstream from the given marker.}
}
\value{
A \code{SInG} object with smoothed read ratios.
}
\description{
Calculate read ratio and smoothed read ratio.
　
Calculate read ratio at each interval based on normalized read counts
obtained via \code{\link[=doSInG]{doSInG()}}. Read ratios calculated at each interval
usually contain data points with stochastic error which are
represented as unexpected changes in read ratio between adjacent
intervals. To remove those errorneous intervals, this function
also calculate average read ratio within the specified window.
}
\examples{
# Load a sample SInG object.
sing_object <- system.file("extdata", "sample.Rdata", package = "SInG")
load(sing_object)

object <- calcRatio(object, window = 5000)

}
