% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSInG.R
\name{doSInG}
\alias{doSInG}
\title{Execute syntenic interval genotyping}
\usage{
doSInG(object, bam_files)
}
\arguments{
\item{object}{A SInG obejct.}

\item{bam_files}{A string vector of paths to BAM files.}
}
\value{
A SInG object with the \code{counts} element which stores a list of
read counts in each interval
}
\description{
Import bam file(s) and count reads in the syntenic intervals.
}
\examples{
# Load a sample SInG object.
sing_object <- system.file("extdata", "sample.Rdata", package = "SInG")
load(sing_object)

# Get path to a sample BAM file.
bam_files <- system.file("extdata", "sample.bam", package = "SInG")

object <- doSInG(object, bam_files)

}
