% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{findFASTQ}
\alias{findFASTQ}
\title{List up FASTQ files to be processed}
\usage{
findFASTQ(in_dir, pattern = NULL, ignore = NULL)
}
\arguments{
\item{in_dir}{A string to indicate the path to the directory storing FASTQ files.}

\item{pattern}{A string to specify the pattern that should be contained in
the FASTQ file names to be returned.}

\item{ignore}{A string to specify the pattern that should not be contained in
the FASTQ file names to be returned.}
}
\value{
A data.frame indicating FASTQ files and sample IDs.
}
\description{
Find FASTQ files in the specified directory and make a list of FASTQ files.
}
\details{
This function searches the files with the .fq or .fastq extension.
Compressed files with the .gz extension can also be detected.  Samples IDs
are automatically obtained by removing characters after _ in the FASTQ file
names. The strings specified to \code{pattern} and \code{ignore} will be used to
specify the \code{pattern} argument in the \code{grep()} function. The FASTQ files
listed by \code{findFASTQ()} should be a set of a single file for each sample,
each file of which is of a single end FASTQ or merged pair end FASTQ files,
or a set of a pair of files for each samples, each file of which is of a pair
end FASTQ file.
}
\examples{
\dontrun{
in_dir <- "path/to/directory/containing/FASTQ_files"
pattern <- "pattern_in_filename_should_be_matched"
ignore <- "pattern_in_filename_should_not_be_matched"
fq_list <- findFASTQ(in_dir, pattern, ignore)
}

}
