% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{findMCPtag}
\alias{findMCPtag}
\title{Find mappable-collinear-polymorphic tags}
\usage{
findMCPtag(ar, dg)

findMCPtag(ar, dg)
}
\arguments{
\item{ar}{An AR class object produced by \code{alignTAG()}.}

\item{dg}{A DG class object produced by \code{digestGenome()}.}
}
\value{
A MCPtag classs object.

A MCPtag classs object.
}
\description{
Find mappable-collinear-polymorphic tags (MCPtags) based on the information
given by the AR and GD class object

Find mappable-collinear-polymorphic tags (MCPtags) based on the information
given by the AR and GD class object
}
\details{
Based on the alignment results of simulated digested genome,
this function list up mappable-collibear-polymorphic tags that are validated
to be used for genotyping a given population derived from a cross between
founders having the reference genome and the alternative genome that were
specified as \code{ref_fn} and \code{alt_fn} in \code{run_mummer()}, \code{digestGenome()}, and
\code{alignTAG()}.

Based on the alignment results of simulated digested genome,
this function list up mappable-collibear-polymorphic tags that are validated
to be used for genotyping a given population derived from a cross between
founders having the reference genome and the alternative genome that were
specified as \code{ref_fn} and \code{alt_fn} in \code{run_mummer()}, \code{digestGenome()}, and
\code{alignTAG()}.
}
\examples{
\dontrun{
mcptag <- findMCPtag(ar = ar, dg = dg)
}

\dontrun{
mcptag <- findMCPtag(ar = ar, dg = dg)
}

}
