% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{makeRE}
\alias{makeRE}
\title{Build a RE object}
\usage{
makeRE(re_names = NULL, re_sites = NULL)
}
\arguments{
\item{re_names}{A character vector to indicate the names of REs.}

\item{re_sites}{A character vector to indicate the RE recognition sites.}
}
\value{
A RE object.
}
\description{
A RE object stores the information of recognition sequences of restriction
enzymes (REs) and also the remnant sequences after the cut by the REs.
}
\details{
The RE object produced via this function is required to run
\code{digestGenome()}. The order of strings specified to \code{re_names} and \code{re_sites}
are important so that the first one indicates the cut site found in the
forward reads (Read 1 of the NGS output) while the second one indicates the
cut site found in the reverse reads (Read 2 of the NGS output).
If not \code{re_sites = NULL}, \code{re_names} will be ignored and
make a RE object using specified \code{re_sites} information. \code{re_sites} requires
a vector of strings in which each string represent the recognition sequence
and the cut position on the top strand. For example, the site of PstI should
be specified as \code{re_sites = "CTGCA/G"} in which the recognition sequence is
"CTGCAG" and the cut site is the position between the 5th A and the 6th G.
The currently supported REs that can be specified to \code{re_names} are listed
below.
\itemize{
\item KpnI
\item PstI
\item MspI
\item ApeKI
}
}
\examples{
\dontrun{
re_names <- c("KpnI", "MspI")
re <- makeRE(re_names)
}

}
