% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{mcptagg}
\alias{mcptagg}
\title{Count reads on MCPtags}
\usage{
mcptagg(
  mcptag,
  fq_list,
  out_dir = "",
  n_threads = 1,
  n_parallel = 2,
  vcf = FALSE,
  bam_list = NULL
)

mcptagg(
  mcptag,
  fq_list,
  out_dir = "",
  n_threads = 1,
  n_parallel = 2,
  vcf = FALSE,
  bam_list = NULL
)
}
\arguments{
\item{mcptag}{A MCPtag class object.}

\item{fq_list}{A character matrix to indicate input FASTQ files. See details.}

\item{out_dir}{A string to specify the output directory.}

\item{n_threads}{A integer to specify the number of threads to run the read
alignment.}

\item{n_parallel}{A integer to specify the number of concurrent sessions to
run read alignments.}

\item{vcf}{A logical value to indicate whether a VCF file should be created
or not. If \code{FALSE}, create only a GDS file.}

\item{bam_list}{A character matrix to indicate BAM files that have already
been created via read alignment.}
}
\value{
A vector of string(s) indicating the path(s) of the output GDS file
and the output VCF file if \code{vcf = TRUE}.

A vector of string(s) indicating the path(s) of the output GDS file
and the output VCF file if \code{vcf = TRUE}.
}
\description{
Align reads on the MCPtags and count the mapped reads.

Align reads on the MCPtags and count the mapped reads.
}
\details{
Align reads by the \code{Rsubread::align()} on the MCPtags prepared by
\code{findMCPtag()} or \code{mergeGenome()}. This function output bam files into
\code{out_dir} and the result of SNP calling and read counts as a GDS file. A GDS
file is the file formatted in the Genomic Data Structure format. You can
access a GDS file using the \code{gdsfmt} package. \code{SeqArray} and \code{SNPRelate} also
provide functions to manipulate a GDS file. \code{GBScleanR} enable you to
estimate true genotypes, haplotypes, and dosages at markers using
an error-prone aware hidden Markov model.

Align reads by the \code{Rsubread::align()} on the MCPtags prepared by
\code{findMCPtag()} or \code{mergeGenome()}. This function output bam files into
\code{out_dir} and the result of SNP calling and read counts as a GDS file. A GDS
file is the file formatted in the Genomic Data Structure format. You can
access a GDS file using the \code{gdsfmt} package. \code{SeqArray} and \code{SNPRelate} also
provide functions to manipulate a GDS file. \code{GBScleanR} enable you to
estimate true genotypes, haplotypes, and dosages at markers using
an error-prone aware hidden Markov model.
}
\examples{
\dontrun{
out_fn <- mcptagg(mcptag = mcptag,
                  fq_list = fq_list,
                  out_dir = "",
                  n_threads = 10, 
                  n_parallel = 5, 
                  vcf = TRUE, 
                  bam_list = NULL)
}
　
\dontrun{
out_fn <- mcptagg(mcptag = mcptag,
                  fq_list = fq_list,
                  out_dir = "",
                  n_threads = 10, 
                  n_parallel = 5, 
                  vcf = TRUE, 
                  bam_list = NULL)
}
　
}
