% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPtaggR-package.R
\docType{package}
\name{MCPtaggR-package}
\alias{MCPtaggR}
\alias{MCPtaggR-package}
\title{MCPtaggR: Mappable-collinear-polymorphic tag genotyping pipeline}
\description{
This package provides functions to perform Mappable-collinear-polymorphic tag genotyping (MCPtagg).
}
\author{
\strong{Maintainer}: Tomoyuki Furuta \email{f.tomoyuki@okayama-u.ac.jp} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
