% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mum2SInG.R
\name{mum2SInG}
\alias{mum2SInG}
\title{Load Mummer's output file#' Load Mummer's output file#' Load Mummer's o#' Load Mummer's output file}
\usage{
mum2SInG(mummer_fn)
}
\arguments{
\item{mummer_fn}{A string to specify the path to output file of
Mummer to load.}
}
\value{
A list with two elements which are \linkS4class{GRanges}
objects storing stretches described in Mummer's output for
reference genome and alternative genome.
}
\description{
Load Mummer's output file#' Load Mummer's output file#' Load Mummer's o#' Load Mummer's output file
}
\examples{
mummer_fn <- system.file("extdata", "sample.mummer", package = "SInG")
object <- mum2SInG(mummer_fn = mummer_fn)

}
