% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{mummer2SNPs}
\alias{mummer2SNPs}
\title{Import the MUMmer output files}
\usage{
mummer2SNPs(mummer_fn)

mummer2SNPs(mummer_fn)
}
\arguments{
\item{mummer_fn}{A vector of two strings indicating a path to.snps file and
a path to .coord file generated by MUMmer.}
}
\value{
A SNPs class object storing available single nucleotide polymorphisms
(SNPs) information.

A SNPs class object storing available single nucleotide polymorphisms
(SNPs) information.
}
\description{
Import .snps and .coord files generated by MUMmer and then apply a filtering
to only retain SNPs in collinear blocks detected between the reference and
the alternative genomes given to MUMmer.

Import .snps and .coord files generated by MUMmer and then apply a filtering
to only retain SNPs in collinear blocks detected between the reference and
the alternative genomes given to MUMmer.
}
\examples{
\dontrun{
snps <- mummer2SNPs(mummer_fn = c("path/to/mummer.snps",
                                  "path/to/mummer.coord"))
}

\dontrun{
snps <- mummer2SNPs(mummer_fn = c("path/to/mummer.snps",
                                  "path/to/mummer.coord"))
}

}
