% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadRatio.R
\name{plotReadRatio}
\alias{plotReadRatio}
\title{Draw line plot of read ratio per interval}
\usage{
plotReadRatio(
  object,
  ind = NULL,
  genome = "ref",
  smooth_only = FALSE,
  coord = c(NA, 2),
  dot_color = "green",
  dot_size = 0.5,
  dot_alpha = 0.3,
  line_color = "magenta",
  line_size = 0.5,
  line_alpha = 0.7,
  font_size = 1,
  ggards = NULL
)
}
\arguments{
\item{object}{A SInG object after running \code{\link[=calcRaio]{calcRaio()}}.}

\item{ind}{An integer vector to specify indices of samples to be
visualized. If multiple integers were specified, lines plots for
the samples will be overlaid with different colors.}

\item{genome}{A string of either "ref" or "alt" to indicate which
interval position information should be used.}

\item{smooth_only}{A logical value to plot only lines of smoothed ratios
or both lines of smoothed ratios and dots of raw ratios.}

\item{coord}{A vector with two integer to specify the number of rows and
columns of tiled line plots for each chromosome.}

\item{dot_color}{A string to specify the color of dots in the plot.}

\item{dot_size}{A numeric value to specify the size of dots in the plot.}

\item{dot_alpha}{A numeric value to specify the level of transparency
of dots in the plot.}

\item{line_color}{A string to specify the color of lines in the plot.
@param line_size A numeric value to specify the width of lines in the plot.}

\item{line_alpha}{A numeric value to specify the level of transparency
of lines in the plot.}

\item{font_size}{A numeric value to specify font size in the plot.}

\item{smooth}{If TRUE, as default, used smoothed read ratio.
Otherwise, raw read ratio.}

\item{ggargs}{A string to modify ggplot.}
}
\value{
Invisibly returns a ggplot object.
}
\description{
Draw line plot(s) of read ratio per interval for samples.
}
\examples{
# Load a sample SInG object.
sing_object <- system.file("extdata", "sample.Rdata", package = "SInG")
load(sing_object)

plotReadRatio(object, ind = 1)

}
