% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{run_mummer}
\alias{run_mummer}
\title{Execute MUMmer}
\usage{
run_mummer(ref_fn, alt_fn, mummer_dir, out_dir = "", out_fn = "mummer_out")

run_mummer(ref_fn, alt_fn, mummer_dir, out_dir = "", out_fn = "mummer_out")
}
\arguments{
\item{ref_fn}{A string to specify the path to the reference genome fasta.}

\item{alt_fn}{A string to specify the path to the alternative genome fasta.}

\item{mummer_dir}{A string to specify the directory in which
the MUMmer executable files including nucmer, delta_filter, show-snps, and
show-coords exist.}

\item{out_dir}{A string to specify the output directory.}

\item{out_fn}{A string to specify the prefix of output files.}
}
\value{
a vector of strings indicating the paths to the output files.

a vector of strings indicating the paths to the output files.
}
\description{
A wrapper function to execute MUMmer from an R environment. This function
requires the path of the MUMmer executable file on your system.

A wrapper function to execute MUMmer from an R environment. This function
requires the path of the MUMmer executable file on your system.
}
\details{
MUMmer is a tool to compare a genome sequence against another
and detect identical genomic region between the given genomes. In addition,
the tool also detect polymorphisms and structural variations. Thus,
\code{run_mummer} requires a reference genome fasta and another to be specified as
\code{ref_fn} and \code{alt_fn}. The alleles found in the reference genome would be listed as
reference alleles in the output vcf file of \code{mcptagg()}, whereas the alleles
found in another specified as \code{alt_fn} would be alternative alleles. Running
MUMmer on repetitive sequences will take longer execution time.
Masking repetitive sequences is highly recommended before running MUMmer.
Please run RepeatMasker or a similar tool on your genomes first, and then
use the masked genome sequences as input to \code{run_mummer()}.

MUMmer is a tool to compare a genome sequence against another
and detect identical genomic region between the given genomes. In addition,
the tool also detect polymorphisms and structural variations. Thus,
\code{run_mummer} requires a reference genome fasta and another to be specified as
\code{ref_fn} and \code{alt_fn}. The alleles found in the reference genome would be listed as
reference alleles in the output vcf file of \code{mcptagg()}, whereas the alleles
found in another specified as \code{alt_fn} would be alternative alleles. Running
MUMmer on repetitive sequences will take longer execution time.
Masking repetitive sequences is highly recommended before running MUMmer.
Please run RepeatMasker or a similar tool on your genomes first, and then
use the masked genome sequences as input to \code{run_mummer()}.
}
\examples{
\dontrun{
mummer_out <- run_mummer(ref_fn = "reference_genome.fa",
                         alt_fn = "alternative_genome.fa",
                         mummer_dir = "path/to/mummer",
                         out_dir = "mummer_out",
                         out_fn = "mummer_result")
}

\dontrun{
mummer_out <- run_mummer(ref_fn = "reference_genome.fa",
                         alt_fn = "alternative_genome.fa",
                         mummer_dir = "path/to/mummer",
                         out_dir = "mummer_out",
                         out_fn = "mummer_result")
}

}
