% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SInG-package.R
\docType{package}
\name{SInG-package}
\alias{SInG}
\alias{SInG-package}
\title{SInG: Syntenic interval genotyping}
\description{
This package provides functions to perform syntenic interval genotyping.
}
\author{
\strong{Maintainer}: Tomoyuki Furuta \email{f.tomoyuki@okayama-u.ac.jp} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
