% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing2VCF.R
\name{sing2VCF}
\alias{sing2VCF}
\title{Write a VCF file.}
\usage{
sing2VCF(object, file_path, dummy_parents = TRUE)
}
\arguments{
\item{object}{A SInG object.}

\item{file_path}{A string to indicate the file path of
the output VCF file.}

\item{dummy_parents}{A logical value to indicate whether insert dummy
data for parental samples or not.}
}
\value{
Invisibly returns TRUE.
}
\description{
Output SInG results in a VCF file with interval position information,
read counts, and read ratio.
}
\examples{
# Load a sample SInG object.
sing_object <- system.file("extdata", "sample.Rdata", package = "SInG")
load(sing_object)

file_path <- tempfile("_SInG", fileext = "vcf")

sing2VCF(object, file_path)

}
