% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{digestGenome}
\alias{digestGenome}
\title{In silico digestion of the input genomes}
\usage{
digestGenome(
  snps,
  ref_fn,
  alt_fn,
  out_dir = "",
  read_len = 150,
  re,
  fragment_len = c(0, 1000),
  reuse = TRUE
)
}
\arguments{
\item{snps}{A SNPs object.}

\item{ref_fn}{A string to specify the path to the reference genome FASTA.}

\item{alt_fn}{A string to specify the path to the alternative genome FASTA.}

\item{out_dir}{A string to specify the output directory.}

\item{read_len}{The maximum length of reads in FASTQ to be mapped on the
genomes in the read mapping step.}

\item{re}{The recognition sequences of restriction enzymes used in your
reduced representation sequencing.}

\item{fragment_len}{A vector of two integers indicating the range of
acceptable fragment lengths.}

\item{reuse}{If TRUE, this function do not generate output files but reuse
the files created by \code{digestGenome()} in the out_dir directory in the
previous execution of \code{digestGenome()}.}
}
\value{
A DG object that is required to run \code{alignTAG()}
}
\description{
Search recognition sites of the specified restriction enzymes in the given
genome sequences and generate FASTA files of the simulated digested fragments.
}
\details{
The \code{re}, \code{read_len}, and \code{fragment_len} arguments should match with the
restriction enzymes, read length, and fragment lengths that were used and
sequenced in your NGS run. If you applied size selection on your sequencing
library during library preparation, \code{fragment_len} should match with the
range of the size selection you did.
}
\examples{
\dontrun{
dg <- digestGenome(snps = snps, 
                   ref_fn = "path/to/ref.fa",
                   alt_fn = "path/to/alt.fa",
                   out_dir = "outdir", 
                   read_len = 150,
                   re = re,
                   fragment_len = c(100, 600), 
                   reuse = FALSE)
}

}
